Extract `replace_region` functions into `view.py`
