+++ core/commands/log_graph.py
@@ -1,5 +1,4 @@
 from collections import deque
-from contextlib import contextmanager, ExitStack
 from functools import lru_cache, partial
 from itertools import chain, islice
 import locale
@@ -27,6 +26,7 @@ from ..runtime import (
     run_or_timeout, run_on_new_thread,
     text_command
 )
+from ..view import replace_region
 from ..ui_mixins.input_panel import show_single_line_input_panel
 from ..ui_mixins.quick_panel import show_branch_panel
 from ...common import util
@@ -223,69 +223,6 @@ def augment_color_scheme(view):
 DATE_FORMAT = 'human'
 FALLBACK_DATE_FORMAT = 'format:%Y-%m-%d %H:%M'
 DATE_FORMAT_STATE = 'trying'
-
-
-@text_command
-def replace_region(view, edit, text, region=None, wrappers=[]):
-    if region is None:
-        # If you "replace" (or expand) directly at the cursor,
-        # the cursor expands into a selection.
-        # This is a common case for an empty view so we take
-        # care of it out of box.
-        region = sublime.Region(0, max(1, view.size()))
-
-    wrappers = wrappers[:] + [stable_viewport]
-    if any(
-        region.contains(s) or region.intersects(s)
-        for s in view.sel()
-    ):
-        wrappers += [restore_cursors]
-
-    with ExitStack() as stack:
-        for wrapper in wrappers:
-            stack.enter_context(wrapper(view))
-        stack.enter_context(writable_view(view))
-        view.replace(edit, region, text)
-
-
-@contextmanager
-def writable_view(view):
-    is_read_only = view.is_read_only()
-    view.set_read_only(False)
-    try:
-        yield
-    finally:
-        view.set_read_only(is_read_only)
-
-
-@contextmanager
-def restore_cursors(view):
-    save_cursors = [
-        (view.rowcol(s.begin()), view.rowcol(s.end()))
-        for s in view.sel()
-    ] or [((0, 0), (0, 0))]
-
-    try:
-        yield
-    finally:
-        view.sel().clear()
-        for (begin, end) in save_cursors:
-            view.sel().add(
-                sublime.Region(view.text_point(*begin), view.text_point(*end))
-            )
-
-
-@contextmanager
-def stable_viewport(view):
-    # Ref: https://github.com/SublimeTextIssues/Core/issues/2560
-    # See https://github.com/jonlabelle/SublimeJsPrettier/pull/171/files
-    # for workaround.
-    vx, vy = view.viewport_position()
-    try:
-        yield
-    finally:
-        view.set_viewport_position((0, 0))  # intentional!
-        view.set_viewport_position((vx, vy))
 
 
 MYPY = False
